requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework==3.15.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
